// src/app/interfaces/caracteristica.ts
export interface CaracteristicaInterface{
    idCharacteristic?:                  number;
    nameCharacteristic:                 string;
    oppositeName:                       string;
    descriptionCharacteristic:          string;
    oppositeDescription:                string;
    descriptionTypeError:               string;
    gradeCharacteristic:                number;
    typeError:                          string;
    dde:                                boolean;
    dii:                                boolean;
    var:                                boolean;
}